+++ src/components/component-templates/SearchBar.html
@@ -73,6 +73,8 @@
             <h3>Available at    McHenry Library (0.20 mi).</h3>
             <OpenLayersMap 
             :open-map="openMap"
+            :locations=locations
+            :loc_num=loc_num
             v-if="openMap"/>
         </div>
       </div>
