+++ src/styles/_openlayersmap.scss
@@ -1,0 +1,7 @@
+@import "variables"; // DO NOT REMOVE THIS LINE OR SCSS WILL NOT WORK!
+.overlay-content {
+    background: #efefef;
+    box-shadow: 0 5px 10px rgb(2 2 2 / 20%);
+    padding: 10px 20px;
+    font-size: 16px;
+}
